% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number corresponding to a state.}

\item{year}{A number describing the year of interest.}
}
\value{
A map plot of a given state visualizing the accidents based on their
location.
}
\description{
The fars_map_state function reads yearly data of FARS datasets,
filters the data to the given state number, and plots the recorded accidents
to their geographic locations based on latitude and longitude data from the
data file. The state number and the years of which you want to plot can be
customized using the \code{state.num} and \code{year} arguments. The function
uses the make_filename and the fars_read functions to read the data sets for
the specified year(s), then filters the data for the given state number, and
plots the accidents to a map of the specific state based on the locations of
accidents. An error will occur if the given state number is invalid.
}
\examples{
\dontrun{
fars_map_state(1, 2019)
fars_map_state(2, 2018)
}

}
