% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A file path to the data file.}
}
\value{
The function returns the data from the csv file as a "data.frame" and
if not saved to a variable name, it prints information about the data set to
the console. The function contains an "if" condition that prints an error if
the data set does not exist in the given location.
}
\description{
The fars_read function reads a Fatality Analysis Reporting
System (FARS) data set that is in csv format. The file to read can be
customized using the \code{filename} argument.
}
\examples{
\dontrun{
fars_read("2020_was_no_accident.csv.bz2")
fars.df <- fars_read("2020_was_no_accident.csv.bz2")
}

}
