% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A number or a vector of numbers describing the year(s) to
analyze.}
}
\value{
A list containing the data tables with two variables ("MONTH",
"year"). Returns NULL in case of an error.
}
\description{
The fars_read_years function gives an overview of how many
accidents have occurred in each month of a given year. You can customize the
year(s) of which you want an overview with the \code{years} argument, which
accepts a single year number or a vector of year numbers. The function reads
the specified FARS data sets and mutates a new variable ("year") for the
certain year, and selects the "MONTH" attribute from the data set, then gives
a list of data tables of these two attributes as a result. An error will
occur if the given year(s) is/are invalid.
}
\examples{
\dontrun{
fars_read_years(2012)
fars_read_years(c(2018, 2019))
fars.nineties <- fars_read_years(1990:2000)
}

}
