% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A number or a vector of numbers describing the year(s) to
analyze.}
}
\value{
A dataframe describing the number of accidents for the given year(s).
}
\description{
The fars_summarize_years function reads yearly data of FARS
data sets and summarizes the number of accidents by month for each year
specified. You can customize the year(s) of which you want a summary with the
\code{years} argument, which accepts a single year number or a vector of year
numbers. This function passes the specified years to the fars_read_years
function, takes the outputted list of data tables and groups the
observations by year and month, and summarizes the number of accidents into a
dataframe.
}
\examples{
\dontrun{
fars_summarize_years(2017)
fars_summarize_years(c(2004, 2006, 2008))
fars.summary <- fars_summarize_years(2019)
}

}
