\name{farsfunction-package}
\alias{farsfunction-package}
\alias{farsfunction}
\docType{package}
\title{
\packageTitle{farsfunction}
}
\description{
\packageDescription{farsfunction}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{farsfunction}
\packageIndices{farsfunction}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{farsfunction}

Maintainer: \packageMaintainer{farsfunction}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
