\name{fars function.}
\alias{farsfunction}
\docType{package}
\title{ Fars Functions}
\description{Example Package}
\details{
The DESCRIPTION file:
\packageDescription{farsfunction}
\packageIndices{farsfunction}
~~ An overview of how to use the package, including the most important functions ~~
}
\keyword{ package }
