% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number or a vector of numbers describing the
year(s).}
}
\value{
A standard file name which looks like 'accident_year.csv.bz2'. This
name is printed when not saved to a variable name.
}
\description{
The make_filename function creates a standard file name in the
form of the string 'accident_', the year, which can be altered via the
\code{year} argument, and the file format (.csv.bz2).
}
\examples{
\dontrun{
make_filename(2020)
make_filename(c(2018, 2019))
new_filename <- make_filename(2017)
}

}
