\name{fars_map_state}
\alias{fars_map_state...}
\docType{function.}
\title{Example fars_map_state}
\description{
create a map of fars events from a specific state.
}
\author{
\packageAuthor{farsfunction}
Maintainer: \packageMaintainer{farsfunction}
}
\keyword{ fars }
\examples{
\dontrun{fars_map_state(20,2013)}
}
