% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string representing a filename}
}
\value{
A data.frame with fars data from the file called 'filename'
}
\description{
This Helper-function checks if the filename given exists and assumes the data
is in .csv.bz2 format, which it will format as data frame.
}
\note{
although the existance of the file (filename) is checked, it is
assumed that the file is in csv format, if not a warning is passed by
readr::read_csv.

this package comes with three example files that can be accessed using
\code{system.file("extdata", "accident_2013.csv.bz2", package = "farsfunctions")}
\code{system.file("extdata", "accident_2014.csv.bz2", package = "farsfunctions")}
\code{system.file("extdata", "accident_2015.csv.bz2", package = "farsfunctions")}
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")}

}
