% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data files of given years}
\usage{ 
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector containing the years of which the
FARS data is asked}
}
\value{
A list of data.frames or NULL if the year resulted in an invalid file
}
\description{
This function returns a list of data files per year, as specified by the
parameter years. The parameter is used to create filenames, using the
make_filename helper function. If an invalid year is passed, the return
value for that year is NULL
}
\note{
if any of the numbers in years is not valid, the list will contain
"NULL" at that index, and a "invalid year" warning will be issued. Much of the
detail in the data available in the FARS dataset is ignored, as only MONTH
and year are selected.
}
\examples{
\dontrun{
fars_read_years(2015)
fars_read_years(c(2013, 2014, 2015))}

}
