% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create FARS data filename by year}
\usage{
make_filename(year)
} 
\arguments{
\item{year}{The year which the datafile is asked for.}
}
\value{
A character string with format: "accident_yyyy.csv.bz2", where
yyyy is the year parameter passed.
}
\description{
This Helper-function constructs a valid FARS filename from a given year
}
\note{
there is no check whether the number is a valid year for the FARS data,
any typo or garbled number will result in a filename, correct or not.
}
\examples{
\dontrun{
make_filename(2015)}
}
